refactor: support deserializing incoming "data" JSON objects
Probably due to an unintended mistake, the Notifications Engine sends
the "data"'s JSON as a quoted string, instead as a JSON object itself.
Our console.redhat.com CloudEvents specification, which in turn adheres
to the Cloud Events specification, defines the "data" key as a JSON
object[1], and therefore here we attempt to parse it as it is
defined in the specification. However, a fallback of the original
parsing way is left too, in order to avoid breaking the integration
until the bug is fixed in the Notifications Engine. More information
about this can be found in the RHCLOUD-24986[2] Jira ticket.

[1]: https://github.com/RedHatInsights/event-schemas/blob/4119fc6e3820c6519f8f87abc27cdb4604a80db0/schemas/events/v1/events.json#L58-L59
[2]: https://issues.redhat.com/browse/RHCLOUD-24986